Move ShaderDemo to its own file
